Bug Fix; Change in interface; Optimizations;
Bug Fix:
Polynomials coefficents will be computed only when needed. Went for a
lazy solution here : User will be responsible of calling update() if it
become necessary (avoid checks in get(), passing it const and permiting
a substantial speed gain).
This led to interface changes (may continue in the futur...) and small
optimizations.